Updated Shooter Subsystems + workspace
I don't like re-committing the workspace, but I did it to be safe.  Most
notibly fixed the issue of not actually "destroying" subsystems' motor
controllers and other WPI things.  Also, added analog shooter angle when
that gets done mechanically.  Various Bug fixes and some commands.  This
is a jumbled commit...