Added Timer to check how long vision processing takes for each frame
The next frame processed by grip may take a while to get here, so there
needs to be a way to time that for testing purposes.  A window is
"stale" if it is more than 200ms old, for the window data will not be
EXACTLY the same every frame.  If it is, then it is probably not updated