Changed MultiSource + added camera to window height
Changed the multisource module to be able to contain multiple outputs
and sources along with the associated requirements for such as system in
the subsystem class.  Updated drive-train to conform to this better.
Changed Bad practice in OzMath to not use hard-coded constant; just
removed entirely and added hypotinuse finder