Added Function for pixel to angle
I added the functions for converting pixels and angle measurements.  The
field of view for the camera was measued poorly, but the data found
online Here:
http://vrguy.blogspot.com/2013/04/converting-diagonal-field-of-view-and.html
and Here: http://buntworthy.github.io/Wecam-field-of-views/  Assuming
the sensor is 4x3 aspect ratio.  Math is tested in WolframAlpha, and the
programming implementation is tested to comply.  Removed this function
from OzMath.