Added Autonomouse Commands (NOT wired to DS)
Added the autonomouse commands along with a command that shoots the ball
and spins down the wheels.  This also works during teleop.  The auto and
spy auto groups are not hooked up to the smart dashboard, but that
should be as simple as a few edit boxes and retrieving it during
auto-init