Updated GRIP file + bugfixes with DT
Changed the drive train to turn CCW with a positive angle, changed grip
file to be more efficient and export HSV binary mask, still needs to be
uploaded to the pi though.  I have also removed useless config values
from DT.  Changed the method of setting setpoint to always be relative
to the current PID setting.  Fixed potential issues with the
"CenterWindowInView" to actually work.